requirements.txt
requests
beautifulsoup4
openai
pydantic
tiktoken
textstat
wikipedia
python-dotenv
